backend/requirement.txt
django
psycopg2
python-dotenv
djangorestframework
